+++ src/inet/linklayer/configurator/L2NetworkConfigurator.h
@@ -55,7 +55,7 @@ class INET_API L2NetworkConfigurator : public cSimpleModule
 
       public:
         Node(cModule *module) : Topology::Node(module->getId()) { this->module = module; interfaceTable = nullptr; }
-        ~Node() { for (int i = 0; i < (int)interfaceInfos.size(); i++) delete interfaceInfos[i]; }
+        ~Node() { for (size_t i = 0; i < interfaceInfos.size(); i++) delete interfaceInfos[i]; }
     };
 
     /**
